changed id to autoinc in regions

+++ src/db/sql/regions/get-by-ids.sql
@@ -5,7 +5,7 @@ from
 where
   (id) in
     (
-      select (c->>'id') from
+      select (c->>'id')::int4 from
         (
           select unnest(${where:raw})::json c
         ) x
