+++ src/db/sql/regions/update.sql
@@ -3,7 +3,7 @@ set ${set:raw}
 where
   (id) in
     (
-      select (c->>'id') from
+      select (c->>'id')::int4 from
         (
           select unnest(${where:raw})::json c
         ) x
